Fixed now dysfunctional back button
Back button wasn't working because of the new routes that could lead to unit stats display. Before the back button was just a link back to the homepage, so if you access unit stats from a faction page, the back button will just go back to the homepage. This was fixed by using the history.back() method.